requirements.txt
grpcio==1.58.0
grpcio-tools==1.58.0
futures==3.0.5
setuptools==65.5.0
rich==13.5.3